Labels for blocking_type axis
